% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_strings.R
\name{cluster_strings_cc}
\alias{cluster_strings_cc}
\title{Cluster Strings via Connected Components}
\usage{
cluster_strings_cc(s_vec, max_dl = 3, method = "osa", clean = T)
}
\arguments{
\item{s_vec}{a vector of character strings}

\item{max_dl}{max distance (typically damerau-levenshtein) between related strings.}

\item{method}{one of "osa","lv","dl" (see stringdist's methods)}

\item{clean}{whether to space-squish and de-duplicate s_vec}
}
\value{
a data frame containing cluster membership for each input string
}
\description{
Cluster Strings via Connected Components
}
\examples{
cluster_strings_cc(c("alcool","alcohol","alcoholic","brandy","brandie","cachaça"))
}
