% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_strings.R
\name{cluster_strings_eb}
\alias{cluster_strings_eb}
\title{Cluster Strings by Edge-Betweeness}
\usage{
cluster_strings_eb(s_vec, max_dl = 3, clean = T, method = "osa")
}
\arguments{
\item{s_vec}{a vector of character strings}

\item{max_dl}{max distance (typically damerau-levenshtein) between related strings.}

\item{clean}{whether to space-squish and de-duplicate s_vec}

\item{method}{one of "osa","lv","dl" (see stringdist's methods)}
}
\value{
a data frame containing cluster membership for each input string
}
\description{
Cluster Strings by Edge-Betweeness
}
\examples{
s_vec <- c("alcool","alcohol","alcoholic","brandy","brandie","cachaça")
cluster_strings_eb(s_vec)
}
