% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quijote_words}
\alias{quijote_words}
\title{Distinct words in Cervantes' "Don Quijote".}
\format{A data frame w/ ~22k rows and 3 cols:
\describe{
  \item{word}{the unique word, in Spanish}
  \item{len}{the word's length}
  \item{freq}{number of appearances in text}
}}
\source{
\url{http://www.gutenberg.org/cache/epub/2000/pg2000.txt}
}
\usage{
quijote_words
}
\description{
Dataframe listing all distinct words (length>3), their length,
and frequency of appearance in text.
}
\keyword{datasets}
